Rename analog_streaming_client.py to qtm_streaming_client.py
