+++ bomi/device_managers/qtm_streaming_client.py
